+++ src/main/java/mods/gregtechmod/init/ClientEventHandler.java
@@ -7,9 +7,9 @@ import mods.gregtechmod.api.GregTechObjectAPI;
 import mods.gregtechmod.api.util.Reference;
 import mods.gregtechmod.compat.ModHandler;
 import mods.gregtechmod.core.GregTechMod;
-import mods.gregtechmod.core.GregTechTEBlock;
 import mods.gregtechmod.model.*;
 import mods.gregtechmod.objects.BlockItems;
+import mods.gregtechmod.objects.GregTechTEBlock;
 import mods.gregtechmod.objects.blocks.teblocks.base.TileEntityIndustrialCentrifugeBase;
 import mods.gregtechmod.objects.items.ItemCellClassic;
 import mods.gregtechmod.util.*;
