+++ src/main/java/mods/gregtechmod/init/RegistryHandler.java
@@ -8,9 +8,9 @@ import mods.gregtechmod.api.cover.ICoverProvider;
 import mods.gregtechmod.api.util.Reference;
 import mods.gregtechmod.compat.ModCompat;
 import mods.gregtechmod.core.GregTechMod;
-import mods.gregtechmod.core.GregTechTEBlock;
-import mods.gregtechmod.cover.Cover;
+import mods.gregtechmod.objects.Cover;
 import mods.gregtechmod.objects.GregTechComponent;
+import mods.gregtechmod.objects.GregTechTEBlock;
 import mods.gregtechmod.objects.blocks.tileentities.TileEntityLightSource;
 import mods.gregtechmod.util.GtUtil;
 import net.minecraft.block.Block;
