+++ src/main/java/mods/gregtechmod/objects/blocks/teblocks/base/TileEntityCoverable.java
@@ -5,10 +5,10 @@ import ic2.core.util.StackUtil;
 import mods.gregtechmod.api.cover.CoverType;
 import mods.gregtechmod.api.cover.ICover;
 import mods.gregtechmod.api.cover.ICoverable;
-import mods.gregtechmod.cover.Cover;
-import mods.gregtechmod.cover.CoverGeneric;
-import mods.gregtechmod.cover.CoverVent;
+import mods.gregtechmod.objects.Cover;
 import mods.gregtechmod.objects.blocks.teblocks.component.CoverHandler;
+import mods.gregtechmod.objects.covers.CoverGeneric;
+import mods.gregtechmod.objects.covers.CoverVent;
 import mods.gregtechmod.util.GtUtil;
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.entity.player.EntityPlayer;
